agregando documentacion para el bloc, sus estados y eventos
