--- lib/presentation/blocs/movies_bloc/movies_event.dart
@@ -1,7 +1,0 @@
-part of 'movies_bloc.dart';
-
-// Eventos
-abstract class MovieEvent {}
-
-class FetchMoviesEvent extends MovieEvent {}
-class FetchMoreMoviesEvent extends MovieEvent {}
