+++ lib/presentation/widgets/widgets.dart
@@ -1,4 +1,3 @@
-export   'package:movie_app_uplabs/presentation/widgets/movies/movie_grid_listview.dart';
+export 'package:movie_app_uplabs/presentation/widgets/movies/movie_grid_listview.dart';
 export 'package:movie_app_uplabs/presentation/widgets/shared/custom_appbar.dart';
 export 'package:movie_app_uplabs/presentation/widgets/shared/custom_bottom_navbar.dart';
-
